Skip CLI tests on Python 2.6
